+++ HW_1/Hometask_2/list.h
@@ -9,7 +9,7 @@ public:
     {
 
     }
-    virtual void addElement(int number) = 0;
+    virtual void addElement(double number) = 0;
     virtual void deleteElement(int number) = 0;
     virtual void printList() = 0;
     int numberOfElements;
